Improved encapsulation
Changed parameters to be added by composition and split code into header files.
