+++ include/BehaviourTypes.h
@@ -1,0 +1,11 @@
+// BehaviourTypes.h
+#pragma once
+
+#include <string>
+
+enum class BehaviourType { Flocking, Pheremone };
+
+struct BehaviourTypeInfo {
+  BehaviourType type;
+  std::string displayName;
+};
